requirements.txt
lightning[extra]
pydicom
nibabel
opencv-python
SimpleITK
SAM-2 @ git+https://github.com/facebookresearch/segment-anything-2.git
torchmetrics